requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
opencv-python>=4.8.0
ruff>=0.1.0
mypy>=1.5.0